refactor(disagg): split windowed capture responsibilities
